Fix LRN testing (#26)

+++ inference-engine/tests/functional/plugin/shared/include/single_layer_tests/lrn.hpp
@@ -19,8 +19,8 @@ namespace LayerTestsDefinitions {
 
 typedef std::tuple<
         double,                        // Alpha
-        size_t,                        // Beta
-        size_t,                        // Bias
+        double,                        // Beta
+        double,                        // Bias
         size_t,                        // Size,
         InferenceEngine::Precision,    // Network precision
         InferenceEngine::SizeVector,   // Input shapes
